+++ src/utils/tests/unit/featureViaURL/index.spec.js
@@ -1,7 +1,7 @@
 import {createGeoJSON, getFeatureIds} from "../../../featureViaURL";
 import {expect} from "chai";
 import sinon from "sinon";
-import {transform} from "@masterportal/masterportalapi/src/crs";
+import crs from "@masterportal/masterportalapi/src/crs";
 import Map from "ol/Map";
 
 describe("featureViaURL", function () {
@@ -34,8 +34,8 @@ describe("featureViaURL", function () {
 
             geoJSON = createGeoJSON(features, geometryType, epsg);
             geoJSON.features.forEach((feature, index) => {
-                expect(feature.geometry.coordinates).to.eql(transform("EPSG:" + epsg, "EPSG:4326", features[index].coordinates));
-                expect(feature.properties.coordLabel).to.eql(transform("EPSG:" + epsg, "EPSG:4326", features[index].coordinates));
+                expect(feature.geometry.coordinates).to.eql(crs.transform("EPSG:" + epsg, "EPSG:4326", features[index].coordinates));
+                expect(feature.properties.coordLabel).to.eql(crs.transform("EPSG:" + epsg, "EPSG:4326", features[index].coordinates));
                 expect(feature.properties.featureLabel).to.equal(features[index].label);
                 expect(feature.geometry.type).to.equal(geometryType);
                 expect(feature.properties.typeLabel).to.equal(geometryType);